HomeServices?

+++ src/services/HomeServices.js
@@ -2,19 +2,25 @@ import axios from 'axios';
 
 const baseURL = 'https://marble.azurewebsites.net/api';
 
-const token = localStorage.getItem('token');
-
-const headers = {
-  Authorization: `Bearer ${token}`,
-  'Content-Type': 'application/json',
-};
-
 const homeService = {
   getItems() {
     return axios.get(`${baseURL}/Home`);
   },
 
   async postItem(body) {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+      console.error('Token not found. User is not authenticated.');
+
+      return Promise.reject(new Error('User not authenticated.'));
+    }
+
+    const headers = {
+      Authorization: `Bearer ${token}`,
+      'Content-Type': 'application/json',
+    };
+
     try {
       const response = await axios.post(`${baseURL}/Home`, body, { headers });
 
@@ -25,6 +31,19 @@ const homeService = {
   },
 
   async putItem(id, body) {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+      console.error('Token not found. User is not authenticated.');
+
+      return Promise.reject(new Error('User not authenticated.'));
+    }
+
+    const headers = {
+      Authorization: `Bearer ${token}`,
+      'Content-Type': 'application/json',
+    };
+
     try {
       const response = await axios.put(`${baseURL}/Home/${id}`, body, {
         headers,
@@ -37,6 +56,19 @@ const homeService = {
   },
 
   async deleteItem(id) {
+    const token = localStorage.getItem('token');
+
+    if (!token) {
+      console.error('Token not found. User is not authenticated.');
+
+      return Promise.reject(new Error('User not authenticated.'));
+    }
+
+    const headers = {
+      Authorization: `Bearer ${token}`,
+      'Content-Type': 'application/json',
+    };
+
     try {
       const response = await axios.delete(`${baseURL}/Home/${id}`, {
         headers,
